+++ tools/cli/Cargo.toml
@@ -19,6 +19,7 @@ diem-config = { workspace = true }
 diem-node = { workspace = true }
 libra-config = { workspace = true }
 libra-framework = { workspace = true }
+libra-genesis-tools = { workspace = true }
 libra-query = { workspace = true }
 libra-txs = { workspace = true }
 libra-types = { workspace = true }
