+++ tools/genesis/src/lib.rs
@@ -1,6 +1,7 @@
-// pub mod convert_types;
-// pub mod ancestry;
+//! genesis
+pub mod cli;
 pub mod compare;
+
 pub mod genesis;
 pub mod genesis_builder;
 pub mod genesis_functions;
