+++ tools/genesis/src/supply.rs
@@ -174,61 +174,3 @@ pub fn populate_supply_stats_from_legacy(rec: &[LegacyRecoveryV6]) -> anyhow::Re
     pb.finish_and_clear();
     Ok(supply)
 }
-
-// #[test]
-// fn test_genesis_math() {
-//     use std::path::PathBuf;
-//     let p = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-//         .join("tests/fixtures/sample_export_recovery.json");
-
-//     let r = crate::parse_json::recovery_file_parse(p).unwrap();
-
-//     // let settings = SupplySettings {
-//     //     target_supply: 100_000_000_000.0, // 100B
-//     //     target_future_uses: 0.70,
-//     //     years_escrow: 7,
-//     //     map_dd_to_slow: vec![
-//     //         // FTW
-//     //         "3A6C51A0B786D644590E8A21591FA8E2"
-//     //             .parse::<LegacyAddress>()
-//     //             .unwrap(),
-//     //         // tip jar
-//     //         "2B0E8325DEA5BE93D856CFDE2D0CBA12"
-//     //             .parse::<LegacyAddress>()
-//     //             .unwrap(),
-//     //     ],
-//     // };
-
-//     // // confirm the supply of normal, slow, and donor directed will add up to 100%``
-
-//     // let mut supply = populate_supply_stats_from_legacy(&r, &settings.map_dd_to_slow).unwrap();
-//     // dbg!(&supply);
-
-//     println!("before");
-//     let pct_normal = supply.normal / supply.total;
-//     let pct_dd = supply.donor_voice / supply.total;
-//     let pct_slow = supply.slow_total / supply.total;
-//     let pct_mk_whole = supply.make_whole / supply.total;
-//     let _pct_val_locked = supply.slow_validator_locked / supply.total;
-
-//     let sum_all_pct = pct_normal + pct_slow + pct_dd + pct_mk_whole;
-//     assert!(sum_all_pct == 1.0);
-//     assert!(supply.total == 2401575768999244.0);
-
-//     // genesis infra escrow math
-//     // future uses is intended to equal 70% in this scenario.
-
-//     supply.set_ratios_from_settings(&settings).unwrap();
-//     dbg!(&supply);
-
-//     // escrow comes out of validator locked only
-//     let to_escrow = supply.escrow_pct * supply.slow_validator_locked;
-//     let new_slow = supply.slow_total - to_escrow;
-//     // dbg!(&pct_normal);
-//     // dbg!(&pct_dd);
-//     // dbg!(new_slow / supply.total);
-//     // dbg!(to_escrow / supply.total);
-
-//     let sum_all = to_escrow + new_slow + supply.normal + supply.donor_voice + supply.make_whole;
-//     assert!(supply.total == sum_all);
-// }
